Split electron-main into smaller chunks
