+++ tsconfig.json
@@ -10,6 +10,7 @@
     "outDir": "./lib",
     "rootDir": "./src",
     "declaration": true,
+    "typeRoots": ["src/@types"],
     "lib": [
       "es2019",
       "dom"
